Refactor cmd into command-focused packages
