+++ cmd/notifications/notifications.go
@@ -1,0 +1,27 @@
+/*
+Copyright © 2025 NAME HERE <EMAIL ADDRESS>
+*/
+package notifications
+
+import (
+	"github.com/spf13/cobra"
+)
+
+var (
+	reason           string
+	repo             string
+	owner            string
+	unseen           bool
+	notificationType string
+)
+
+// notificationsCmd represents the notifications command
+var notificationsCmd = &cobra.Command{
+	Use:   "notifications",
+	Short: "List notifications",
+	Long:  `List notifications from GitHub.`,
+}
+
+func Cmd() *cobra.Command {
+	return notificationsCmd
+}
