+++ cmd/notifications/result.go
@@ -1,4 +1,4 @@
-package cmd
+package notifications
 
 import (
 	"fmt"
